requirements.txt
flask
gunicorn
faiss-cpu
langchain
langchain-community
langchain-huggingface
sentence-transformers
transformers
requests




